pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ethical-jurisprudence-core"
version = "1.3.0"
description = "Ethical Jurisprudence Core (EJC) - Rights-Based Jurisprudence Architecture for the Mutual Intelligence Framework"
readme = "README.md"
authors = [
    {name = "William Parris", email = "will@eleanorproject.org"}
]
license = {text = "CC BY 4.0"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "sqlalchemy>=2.0",
    "flask>=3.0",
    "sentence-transformers>=2.2",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "tenacity>=8.2",
    "openai>=1.0",
    "anthropic>=0.18",
    "google-generativeai>=0.3",
    "pydantic-settings>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-xdist>=3.3",
    "black>=23.7",
    "ruff>=0.0.286",
    "mypy>=1.5",
    "isort>=5.12",
    "bandit>=1.7",
]
docs = [
    "mkdocs>=1.5",
    "mkdocs-material>=9.2",
]
ops = [
    "streamlit>=1.30.0",
    "pandas>=2.1.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
]

[project.urls]
Homepage = "https://github.com/eleanor-project/EJE"
Documentation = "https://eje.readthedocs.io"
Repository = "https://github.com/eleanor-project/EJE"
Issues = "https://github.com/eleanor-project/EJE/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["ejc*", "eje*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
